Fix: 휴대폰 인증 후 회원 존재여부 리턴
