+++ src/main/java/com/example/travelhana/Controller/ExchangeController.java
@@ -1,7 +1,6 @@
 package com.example.travelhana.Controller;
 
 import com.example.travelhana.Dto.ExchangeRequestDto;
-import com.example.travelhana.Exception.Response.ApiResponse;
 import com.example.travelhana.Service.ExchangeService;
 import io.swagger.models.Response;
 import lombok.RequiredArgsConstructor;
