+++ src/main/java/com/example/travelhana/Dto/CodeRequestDto.java
@@ -9,7 +9,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Getter
 @Builder
-public class CodeDto {
+public class CodeRequestDto {
 
+    private String phonenum;
     private String code;
 }
