+++ src/main/java/com/example/travelhana/Dto/CodeResponseDto.java
@@ -1,0 +1,17 @@
+package com.example.travelhana.Dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Builder
+public class CodeResponseDto {
+
+    private Boolean isExistUser;
+    private Boolean isCodeEqual;
+
+}
