+++ src/main/java/com/example/travelhana/Repository/UserRepository.java
@@ -11,6 +11,7 @@ public interface UserRepository extends JpaRepository<User,Integer> {
     Optional<User> findByDeviceId(String deviceId);
     boolean existsByDeviceId(String deviceId);
     Optional<User> findByIdAndIsWithdrawal(int userId, Boolean isWithdrawal);
+    Optional<User> findByPhoneNum(String phonenum);
 
 
 }
