+++ src/main/java/com/example/travelhana/Service/Implement/UserServiceImpl.java
@@ -23,6 +23,7 @@ import com.example.travelhana.Service.UserService;
 import com.example.travelhana.Util.SaltUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -126,7 +127,7 @@ public class UserServiceImpl implements UserService, UserDetailsService {
             ErrorResponse errorResponse=ErrorResponse.builder()
                     .errorMessage(e.getMessage())
                     .build();
-            return ResponseEntity.ok(errorResponse);
+            return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
         }
 
     }
